# DOCS/contracts/openapi.json
{
...
  "openapi": "3.0.1",
  "info": {
    "title": "ConfluxPay API",
    "description": "ConfluxPay backend API â€” multi-tenant payment platform. All requests require a tenant API key carried in the Authorization header.",
    "contact": {
      "name": "ConfluxPay Engineering",
      "email": "engineering@conflux.com"
    },
    "license": {
      "name": "Proprietary",
      "url": "https://conflux.com/license"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8089",
      "description": "Generated server url"
    }
  ],
  "security": [
    {
      "ApiKeyAuth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/merchant/journal": {
      "get": {
        "tags": [
          "merchant-ledger-controller-impl"
        ],
        "operationId": "listJournal",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "occurredAt"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "paginate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListJournalEntryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/merchant/balance": {
      "get": {
        "tags": [
          "merchant-ledger-controller-impl"
        ],
        "operationId": "getBalance",
        "parameters": [
          {
            "name": "code",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "MERCHANT_PAYABLE"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "BDT"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultBalanceDto"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/ledger/journal": {
      "get": {
        "tags": [
          "admin-ledger-controller-impl"
        ],
        "operationId": "listJournal_1",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          },
          {
            "name": "sortBy",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "occurredAt"
            }
          },
          {
            "name": "order",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DESC",
              "enum": [
                "ASC",
                "DESC"
              ]
            }
          },
          {
            "name": "paginate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true
            }
          },
          {
            "name": "sourceType",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "startDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "endDate",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time"
            }
          },
          {
            "name": "ownerId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultListJournalEntryDto"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/ledger/balance/{code}": {
      "get": {
        "tags": [
          "admin-ledger-controller-impl"
        ],
        "operationId": "getBalance_1",
        "parameters": [
          {
            "name": "code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "currency",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "BDT"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResultBalanceDto"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResultMeta": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          },
          "errorCode": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "FORBIDDEN",
              "RESOURCE_NOT_FOUND",
              "RESOURCE_ALREADY_EXISTS",
              "INVALID_OPERATION_STATE",
              "INSUFFICIENT_FUNDS",
              "IDEMPOTENCY_CONFLICT",
              "MFS_ADAPTER_FAILURE",
              "VENDOR_DOWN",
              "QUOTA_EXCEEDED",
              "RISK_REJECTED",
              "TOKEN_EXPIRED",
              "INTERNAL_ERROR"
            ]
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
      "PaginationInfo": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          },
          "totalElements": {
            "type": "integer",
            "format": "int64"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ApiResultListJournalEntryDto": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/JournalEntryDto"
            }
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "JournalEntryDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "sourceType": {
            "type": "string"
          },
          "sourceId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "occurredAt": {
            "type": "string",
            "format": "date-time"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "postings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PostingDto"
            }
          }
        }
      },
      "PostingDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "accountCode": {
            "type": "string"
          },
          "accountType": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          },
          "type": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          }
        }
      },
      "ApiResultBalanceDto": {
        "type": "object",
        "properties": {
          "data": {
            "$ref": "#/components/schemas/BalanceDto"
          },
          "meta": {
            "$ref": "#/components/schemas/ApiResultMeta"
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "BalanceDto": {
        "type": "object",
        "properties": {
          "accountId": {
            "type": "string",
            "format": "uuid"
          },
          "accountCode": {
            "type": "string"
          },
          "accountType": {
            "type": "string"
          },
          "currency": {
            "type": "string"
          },
          "amount": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "ApiKeyAuth": {
        "type": "http",
        "description": "Tenant API key. Send as `Authorization: Bearer <api-key>`. Keys are issued per tenant in Phase 1; rotation invalidates previous keys immediately.",
        "name": "Authorization",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "API_KEY"
      }
    }
  }
}